Refactor type imports to use new types file for Direction
