+++ src/types/types.ts
@@ -1,0 +1,1 @@
+export type Direction = "row" | "column";
